requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.1
Django==4.2
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.1
idna==3.9
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
simplejson==3.19.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
